+++ nautilus_trader/adapters/databento/__init__.py
@@ -12,13 +12,28 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
+"""
+Databento market data integration adapter.
 
+This subpackage provides a data client factory, instrument provider,
+constants, configurations, and data loaders for connecting to and
+interacting with the Databento API, and decoding Databento Binary
+Encoding (DBN) format data.
+
+For convenience, the most commonly used symbols are re-exported at the
+subpackage's top level, so downstream code can simply import from
+``nautilus_trader.adapters.databento``.
+
+"""
 from nautilus_trader.adapters.databento.config import DatabentoDataClientConfig
 from nautilus_trader.adapters.databento.constants import ALL_SYMBOLS
 from nautilus_trader.adapters.databento.constants import DATABENTO
 from nautilus_trader.adapters.databento.constants import DATABENTO_CLIENT_ID
 from nautilus_trader.adapters.databento.factories import DatabentoLiveDataClientFactory
+from nautilus_trader.adapters.databento.factories import get_cached_databento_http_client
+from nautilus_trader.adapters.databento.factories import get_cached_databento_instrument_provider
 from nautilus_trader.adapters.databento.loaders import DatabentoDataLoader
+from nautilus_trader.adapters.databento.providers import DatabentoInstrumentProvider
 from nautilus_trader.core.nautilus_pyo3 import DatabentoImbalance
 from nautilus_trader.core.nautilus_pyo3 import DatabentoStatistics
 
@@ -30,6 +45,9 @@ __all__ = [
     "DatabentoDataClientConfig",
     "DatabentoDataLoader",
     "DatabentoImbalance",
+    "DatabentoInstrumentProvider",
     "DatabentoLiveDataClientFactory",
     "DatabentoStatistics",
+    "get_cached_databento_http_client",
+    "get_cached_databento_instrument_provider",
 ]